Refactor: Refatora o Controller para incluir o ResponseEntity nas minhas rotas

+++ src/main/java/com/example/demo/Ninja/controller/NinjaController.java
@@ -4,6 +4,8 @@ import com.example.demo.Ninja.dto.NinjaDTO;
 import com.example.demo.Ninja.model.NinjaModel;
 import com.example.demo.Ninja.service.NinjaService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,34 +17,55 @@ public class NinjaController {
     @Autowired
     private NinjaService ninjaService;
 
-    @GetMapping (value = "/{id}")
+    @GetMapping(value = "/{id}")
 
-        public NinjaDTO findByIdNinja(@PathVariable Long id){
-        return ninjaService.findByIdNinja(id);
+    public ResponseEntity<?> findByIdNinja(@PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.findByIdNinja(id);
+
+        if(ninja != null){
+            return ResponseEntity.ok(ninja);
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("id: "+ id + " Não encontrado");
         }
-
+    }
 
     @GetMapping()
-    public List<NinjaDTO> FindAll() {
-        return ninjaService.findAllNinja();
+    public ResponseEntity<List<NinjaDTO> >FindAll() {
+        List<NinjaDTO> ninjas = ninjaService.findAllNinja();
+        return ResponseEntity.ok(ninjas);
     }
 
     @PostMapping("/criar")
-    public NinjaDTO create(@RequestBody NinjaDTO ninja){
-         return ninjaService.createNinja(ninja);
+    public ResponseEntity<String> create(@RequestBody NinjaDTO ninja) {
+        NinjaDTO ninjaDTO = ninjaService.createNinja(ninja);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body("Criado com sucesso: " + ninjaDTO.getNome() + "(ID): " + ninjaDTO.getId());
 
     }
 
     @DeleteMapping("/{id}")
-    public void deleteNinja(@PathVariable Long id){
-        ninjaService.deleteNinja(id);
+    public ResponseEntity<String> deleteNinja(@PathVariable Long id) {
+
+        if (ninjaService.findByIdNinja(id) != null) {
+            ninjaService.deleteNinja(id);
+            return ResponseEntity.ok().body("Ninja com id: " + id + " deletado com sucesso");
+
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Id " + id + " não encontrado!");
+        }
     }
 
 
     @PutMapping("/{id}")
-    public NinjaDTO updateNinjaPorId(@PathVariable Long id, @RequestBody NinjaDTO ninja){
-        return ninjaService.updateNinja(id, ninja);
+    public ResponseEntity<?> updateNinjaPorId(@PathVariable Long id, @RequestBody NinjaDTO ninja) {
+       NinjaDTO ninjaAtualizado = ninjaService.updateNinja(id, ninja);
+        if (ninjaAtualizado != null) {
+            return ResponseEntity.ok().body(ninjaAtualizado);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Id " + id + " não encontrado!");
+        }
     }
+
 
 
 }
